requirements.txt
streamlit
streamlit-extras
google-generativeai
python-dotenv
pdf2image
PyPDF2
pytest
